ai-engine/requirements.txt
kafka-python==2.0.2
redis==5.0.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
python-json-logger==2.0.7
xgboost==2.0.3
